Handle timing ourselves
