+++ package.json
@@ -14,7 +14,8 @@
   "license": "MIT",
   "dependencies": {
     "discord.js": "^14.16.2",
-    "dotenv": "^16.4.5"
+    "dotenv": "^16.4.5",
+    "node-cron": "^3.0.3"
   },
   "devDependencies": {
     "prettier": "^3.3.3"
